Detect current page video id more robustly
We previously looked at the URL, but depending on how you click
through from the home screen, the URL will not always show the
correct video ID. So now we pull it from an HTML attribute, which
is surely subject to change but works better to now.
Bumping to new version also.