download srt: use netflix API to construct filename
Recently, Netflix stopped making the name of the current video easily
available, resulting in every downloaded file using the video ID as a
filename.

Luckily we can construct a filename using the Netflix API and the video
ID. This has the benefit of being consistent across all languages and
types of media, though this might be a downside if a particular language
has a very different way of representing show seasons and episodes.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>